V0.3.1 rc (release for addressing CRAN errors) (#115)
* Update create_credentials.R

It is thought that including the only function from the `getPass` pkg in an argument leads some CRAN check machines to believe the Import is unused. This change is thus just for passing R CMD check more widely.

* Limit testing of rmarkdown-based tests

To only those systems with Pandoc 2

* Update Version and NEWS.md

* Added SystemRequirements (pandoc)

This is required according to WRE.

* Substitute function used for checking Pandoc status

Now using the exported `rmarkdown::pandoc_available()` function.