Handle large scale input
The new_releases_email.R test case was failing with a stack
overflow. It was due to backtracking across the contents of
a tag body, which in this case was dozens of megabytes (the
image was 24MB(!!) to start with, then base64-encoded!).

Dropping the (/?) from the pattern allows us to change the
preceding *? to *+ (possessive quantifier, TIL!), which
avoids backtracking.